Split vm.go into vm.go and vm_statusItem.go
